% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{binarize}
\alias{binarize}
\title{Binarize}
\usage{
binarize(img, quantile = 0.95, G = 1L, ...)
}
\arguments{
\item{img}{an imager::cimg}

\item{quantile}{a real, the quantile level used for thresholding.}

\item{G}{an integer, the number of mixture components.}

\item{...}{mclust::densityMclust parameters}
}
\value{
an imager::cimg
}
\description{
Binarize a grayscale image.
}
