% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{binaryLoad}
\alias{binaryLoad}
\title{Load and binarize image}
\usage{
binaryLoad(img_path, threshold)
}
\description{
Load image from given path, negative filter, illumination correction filter, binarization based on threshold
}
