% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{bodyLength}
\alias{bodyLength}
\title{Body length from body image and elongation angle}
\usage{
bodyLength(body_img, elong_angle, viz = FALSE, return_ext = FALSE)
}
\arguments{
\item{body_img}{body image}

\item{elong_angle}{numerical for elongation angle in rad}

\item{viz}{logical. visualization option}

\item{return_ext}{logical. output format, if TRUE, also returns points between which length was measured}
}
\description{
Body length as distance between intersections of body outline with line passing
through the centroid of the body and minimizing lateral point density difference
along body elongation axis defined using PCA
}
