% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{bodyLengthPC}
\alias{bodyLengthPC}
\title{Fast body length from body points}
\usage{
bodyLengthPC(body_points)
}
\arguments{
\item{body_points}{a dataframe or matrix of body points coordinates}
}
\description{
Fast body length measurement using max x values of PCA rotation
}
