% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{bool_first}
\alias{bool_first}
\title{Accessory logical for boolSeqLim}
\usage{
bool_first(x0, x1)
}
\arguments{
\item{x0}{logical vector}

\item{x1}{logical vector}
}
\description{
Accessory logical for boolSeqLim
}
\keyword{internal}
