% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{boolSeqLim}
\alias{boolSeqLim}
\title{Interval of all True/1 sequence in boolean vector}
\usage{
boolSeqLim(v, circular = T)
}
\arguments{
\item{v}{a vector of logical values}

\item{circular}{a logical for prior assumption of circularity of v}
}
\description{
Give matching index of first and last True values of sequences of True in v
}
