% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{centerContour}
\alias{centerContour}
\title{Center Contour}
\usage{
centerContour(outline)
}
\arguments{
\item{outline}{matrix or dataframe of points coordinates}
}
\description{
Center baserd on outline centroid
}
\keyword{internal}
