% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUtils.R
\name{channel_hist}
\alias{channel_hist}
\title{Channel histogram}
\usage{
channel_hist(img, ch = c("R", "G", "B"))
}
\arguments{
\item{img}{an imager::cimg}

\item{ch}{a character vector, subset of ("R", "G", "B").}
}
\value{
a ggplot2 graph
}
\description{
Histogram ggplot of each color channel of an image.
}
