% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{checkSym}
\alias{checkSym}
\title{Check point cloud symmetry along an angle}
\usage{
checkSym(pts, angle = 0, cuts = 10)
}
\arguments{
\item{pts}{binary object as point cloud of even density}

\item{angle}{angle in rad}

\item{cuts}{number of bins to create along x}
}
\description{
Bilateral asymmetry scoring for a point cloud of constant density using
ratio of points above centroid in multiple bins along x-axis (after rotation using angle)
}
