% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{cleanBodyShape}
\alias{cleanBodyShape}
\title{Clean body shape as cimg using morphological closing following by opening}
\usage{
cleanBodyShape(body_img, kernel_size = 3)
}
\arguments{
\item{body_img}{body img cimg.}

\item{kernel_size}{numerical integer. closing and opening kernel size in pixels}
}
\description{
Square closing then closing of size kernel_size on binary body points body_lab_points
Returns a cimg image
}
