% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUtils.R
\name{cm_to_pixel}
\alias{cm_to_pixel}
\title{Centimeter to pixel}
\usage{
cm_to_pixel(ruler, quantile = 0.005)
}
\arguments{
\item{ruler}{an imager::pixset}

\item{quantile}{a real, the quantile level used for thresholding.}
}
\value{
a data.frame
}
\description{
Get the centimeter to pixel conversion factor from a ruler defined as
a pixset.
}
