% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUtils.R
\name{combine}
\alias{combine}
\title{Combine pixset}
\usage{
combine(pixset_ref, pixset_list, padding = 50L)
}
\arguments{
\item{pixset_ref}{an imager::pixset, the reference.}

\item{pixset_list}{a list of imager::pixset, the list of pixsets to combine
with the reference.}

\item{padding}{an integer, the stencil length}
}
\value{
a vector of boolean
}
\description{
Combine pixsets into a reference pixset according to a shared neighbourhood.
The neighbourhood is defined as a squared stencil centered around each
candidate pixset.
}
