% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{contourAngles}
\alias{contourAngles}
\title{Contour angle for each point given a search window = index of the next value to check}
\usage{
contourAngles(coords, search_w = 5)
}
\description{
Contour angle for each point given a search window = index of the next value to check
}
