% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{contourTurns}
\alias{contourTurns}
\title{Contour angle peaks, convex or concave}
\usage{
contourTurns(
  coords,
  search_w = 5,
  splines_df = 30,
  angle_thresh = 0.15,
  viz = 0
)
}
\description{
Modified from fxLPJ.R / Peaks of angle contour for a givens search window, smoothed using splines
}
