% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{coordsAsImg}
\alias{coordsAsImg}
\title{Convert xy coordinates dataset as square c.img with padding}
\usage{
coordsAsImg(coords, padding = 0, return_offset = F)
}
\description{
Convert xy coordinates dataset as square c.img with padding
}
