% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{coordsToLinear}
\alias{coordsToLinear}
\title{Convert xy coordinates as linear index in given c.img}
\usage{
coordsToLinear(coords, img)
}
\description{
Convert xy coordinates as linear index in given c.img
}
