% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{correct_illumination}
\alias{correct_illumination}
\title{Correct illumination}
\usage{
correct_illumination(img, nsamples = 10000L)
}
\arguments{
\item{img}{an imager::cimg}

\item{nsamples}{an integer, pixel subsampling value.}
}
\value{
an imager::cimg object
}
\description{
Correct a gray-scaled image illumination by fitting a linear model and
removing the spatial trend.
}
