% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{create_dir}
\alias{create_dir}
\title{Create directory}
\usage{
create_dir(name, msg = TRUE)
}
\arguments{
\item{name}{a string. name and path of the directory to create}

\item{msg}{logical. console message option}
}
\description{
Create directory if it doesn't already exists
}
\keyword{internal}
