% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{cropPoints}
\alias{cropPoints}
\title{Cropping points list}
\usage{
cropPoints(list_points, list_size, scaling_factor = 1)
}
\arguments{
\item{list_points}{A list of numeric vectors, each representing a point (typically of length 2: x and y).}

\item{list_size}{A numeric vector or list of same length as \code{list_points}, representing the diameter or window size around each point.}

\item{scaling_factor}{A numeric scalar used to scale the size of each bounding box (default is 1).}
}
\value{
A list of numeric vectors of length 4 (xmin, xmax, ymin, ymax) representing the cropped bounding boxes.
}
\description{
Accessory function for gCrop,
}
