% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUtils.R
\name{cyclic_distance}
\alias{cyclic_distance}
\title{Function to calculate cyclic distance between two indices}
\usage{
cyclic_distance(index1, index2, length)
}
\description{
Function to calculate cyclic distance between two indices
}
