% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{desaturation}
\alias{desaturation}
\title{Reduce saturation}
\usage{
desaturation(img, ratio = 2L)
}
\arguments{
\item{img}{an imager::cimg}

\item{ratio}{an integer, how much to divide the saturation by.}
}
\value{
an imager::cimg
}
\description{
Reduce the saturation of an image through HSV conversion.
}
