% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{diamondKern}
\alias{diamondKern}
\title{Make diamond kernel matrix}
\usage{
diamondKern(sz, msg = TRUE)
}
\arguments{
\item{sz}{kernel diameter size}

\item{msg}{boolean. Show rounding message}
}
\description{
Make diamond kernel matrix
}
