% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{dilBodies}
\alias{dilBodies}
\title{Body dilation proportionnal to body length}
\usage{
dilBodies(body_img, body_length)
}
\arguments{
\item{body_img}{Binary pixset or c.img of waterstrider body}

\item{body_length}{A numerical for length of corresponding body}
}
\description{
Body dilation proportionnal to body length
}
