% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUtils.R
\name{ehull}
\alias{ehull}
\title{Ellipsoid hull}
\usage{
ehull(pixset, size = 6L)
}
\arguments{
\item{pixset}{an imager::pixset}

\item{size}{an integer, the morphological opening factor.}
}
\value{
a data.frame
}
\description{
Fit an ellipsoid hull for a pixset with optional morphological opening.
}
