% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{elongAngle}
\alias{elongAngle}
\title{Elongation angle of point cloud maximizing points symmetry}
\usage{
elongAngle(pts)
}
\arguments{
\item{pts}{matrix or dataframe of 2D points of an object with even density}
}
\description{
Elongation angle of point cloud maximizing points symmetry
}
