% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{enhance_contrast}
\alias{enhance_contrast}
\title{Enhance contrast}
\usage{
enhance_contrast(img)
}
\arguments{
\item{img}{an imager::cimg}
}
\value{
an imager::cimg
}
\description{
Enhance the contrasts of an image by running an histogram equalization
separately on each channel and combining the results.
}
