% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{f_prime}
\alias{f_prime}
\title{Accesory function for GMMdRoots : creates derivative function of GMM}
\usage{
f_prime(x, GMM_df)
}
\description{
Accesory function for GMMdRoots : creates derivative function of GMM
}
