% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{find_closest_sequences}
\alias{find_closest_sequences}
\title{find the closest sequences in both directions}
\usage{
find_closest_sequences(rear_index, true_sequences, vector_length)
}
\description{
find the closest sequences in both directions
}
