% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{find_true_sequences}
\alias{find_true_sequences}
\title{Identify sequences of TRUE values in a cyclic boolean vector}
\usage{
find_true_sequences(boolean_vector)
}
\description{
Identify sequences of TRUE values in a cyclic boolean vector
}
