% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gModels.R
\name{formatBioLDA}
\alias{formatBioLDA}
\title{Make sex-wing LDA output understandable}
\usage{
formatBioLDA(pred, angle)
}
\arguments{
\item{pred}{output of predict() on WaterStrideR's LDA model}

\item{angle}{Vector of orientation angle of individuals.
Number of non-NA values should match the number of predictions.}
}
\description{
Separates sex and wing probability from raw LDA output.
}
