% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gBody}
\alias{gBody}
\title{Body segmentation}
\usage{
gBody(img, centroid, k_size, viz = TRUE, a = 1, s = 1, msg = TRUE)
}
\arguments{
\item{img}{input image as pixset or c.img}

\item{centroid}{coordinates of body centroid}

\item{k_size}{diamond kernel size to close gaps between Canny edges}

\item{viz}{visualization option for contouring and hole filling}

\item{a}{alpha parameter for imager::cannyEdges() (threshold adjusment
factor)}

\item{s}{sigma parameter for imager::cannyEdges() (smoothing)}

\item{msg}{return warnings.}
}
\description{
Get precise and threshold-independant segmentation from rough body detection
using Canny edges and kernel flood filling
}
