% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gModels.R
\name{gBodyPredict}
\alias{gBodyPredict}
\title{Prediction pipeline for sex or wings}
\usage{
gBodyPredict(img_data, angle, what)
}
\arguments{
\item{img_data}{c.img or list of c.img of segmented bodies}

\item{angle}{angle, vector of angles or list of angles of the segmented bodies}

\item{what}{"sex" or "wing". Data to be predicted.}
}
\description{
Wrapper for gBodyPredict1, allowing for format control and vectorization
From segmented body binary images: contour, orientation, resampling,
}
