% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_models.R
\name{gBodyPredict1}
\alias{gBodyPredict1}
\title{Prediction pipeline for sex or wings - single image, to be used in gBodyPredict}
\usage{
gBodyPredict1(img, ang, rs, pca, lda)
}
\arguments{
\item{img}{a c.img object. Segmented binary waterstrider body image.}

\item{ang}{a number. Orientation angle of that individual.}

\item{rs}{an integer.}

\item{pca}{a prcomp object.}

\item{lda}{a MASS lda object.}
}
\description{
Prediction pipeline for sex or wings - single image, to be used in gBodyPredict
}
