% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gCleanBin}
\alias{gCleanBin}
\title{Clean individuals binary image}
\usage{
gCleanBin(l_img_bin, centroid, as_coords = TRUE, px_filter = 25)
}
\arguments{
\item{l_img_bin}{An image as pixset or c.img or a list of images}

\item{centroid}{Numeric xy coordinates of the body centroid, or a list body centroids}

\item{as_coords}{A boolean}

\item{px_filter}{A numerical. Minimum amount pixel to keep label}
}
\description{
Removes components touching the edges except for the body + removes small spots
}
