% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gConnectLeg}
\alias{gConnectLeg}
\title{Connect leg landmarks with body}
\usage{
gConnectLeg(body, landmarks, viz = F)
}
\description{
As gLegLandmarks finds insertion point on dilated body, the insertion landmark is biased
Given body as xy coords and landmarks as output of gLegLandmarks, corrects the insertion point
}
