% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gCrop}
\alias{gCrop}
\title{Crop individual gerris}
\usage{
gCrop(base_img, centroids, sizes, viz = F, factor = 3.5)
}
\description{
Square crop around individual gerris centroids from image base_img with a diameter of factor*size
Base image must be imported with imager package, centroids and sizes must be lists of same length
Returns a list of inverted grayscale cropped images
}
