% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gModels.R
\name{gDirection}
\alias{gDirection}
\title{Predict direction of head of Microvelia longipes}
\usage{
gDirection(body, dil_cont, inter_idx, el_angle, viz = TRUE)
}
\arguments{
\item{body}{c.img. Segmented body.}

\item{dil_cont}{numerical matrix of data frame. Dilated body contour.}

\item{inter_idx}{boolean vector matching dil_cont rows. Limb intersections
with dilated body contour.}

\item{el_angle}{numerical in radians. Body elongation angle.
confidence threshold under which returned angle should be NA.}

\item{viz}{boolean. Visualization option.}
}
\description{
Uses body elongation axis, limb-body intersections and body EFA coefficients
to predict direction of head as angle from centroid using trained LDA model.
}
