% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{get_centroid}
\alias{get_centroid}
\title{Get centroid of a pixset.}
\usage{
get_centroid(pixset)
}
\arguments{
\item{pixset}{an imager::pixset}
}
\value{
an integer vector
}
\description{
Get centroid of a pixset.
}
