% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{getScale}
\alias{getScale}
\title{Manual scaling of an image}
\usage{
getScale(img, scale_value = NA, plot_col = "#ff9100", lwd = 1, cex_max = 1.2)
}
\arguments{
\item{img}{a c.img or pixset}

\item{scale_value}{scale value. If NA: value will be asked in console}

\item{plot_col}{string or numerical for plot color}

\item{lwd}{numerical. line width on graphic windows between scale points.}

\item{cex_max}{numerical. black point size of scale points.}
}
\value{
number of pixels per unit
}
\description{
Scale imager image or pixset by setting scale value for a distance between
two pixels set manually on image
}
