% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gGMMThresh}
\alias{gGMMThresh}
\title{Binarize gerris using GMM derivative}
\usage{
gGMMThresh(gerris_crops, nG = 2, noise_n = 0.005, y_root = -25, msg = T)
}
\description{
Binarize gerris by setting a fixed threshold per input image using GMM derivative
}
