% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gLegLandmarks}
\alias{gLegLandmarks}
\title{Find knee and ankle of gerris leg}
\usage{
gLegLandmarks(
  leg_coords,
  insertion,
  inser_thresh = 0.1,
  tresh_ankle = 0.12,
  viz = F,
  msg = T
)
}
\description{
Ankle and knee position using angle along the contour of the leg,
oriented with the body insertion point
}
