% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gLegLandmarksLoop}
\alias{gLegLandmarksLoop}
\title{Loop logic for leg landmarking in gPipeline}
\usage{
gLegLandmarksLoop(leg_coords, insertion, ...)
}
\arguments{
\item{leg_coords}{numerical matrix or dataframe or list of those types.
leg points coordinates}

\item{insertion}{numerical matrix or dataframe or list of those types.
insertion points coordinates}

\item{...}{arguments to be passed to gLegLandmarks}
}
\description{
Loop logic for leg landmarking in gPipeline
}
