% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gLegSeg}
\alias{gLegSeg}
\title{Leg segmentation}
\usage{
gLegSeg(gerris, dilated_body, intersection_coords, insertions)
}
\arguments{
\item{gerris}{\code{cimg} or list of \code{cimg}}

\item{dilated_body}{\code{cimg} or list of \code{cimg}}

\item{intersection_coords}{\verb{x,y} matrix or list of \verb{x,y} matrices}

\item{insertions}{\verb{x,y} matrix or list of \verb{x,y} matrices}
}
\description{
Leg segmentation using insertion points as output of gLegInsertion to detect matching limbs
first separates limbs by substracting dilated body to full individual
}
