% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gLimbInter}
\alias{gLimbInter}
\title{Gerris limbs intersection points}
\usage{
gLimbInter(v)
}
\arguments{
\item{v}{a vector of logical values}
}
\description{
For v a boolean of intersection points in the contour, assume limb intersections as continuous sequences
of intersection values in the contour. Compute first, last, and middle point of each sequence
}
