% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gMeasureLeg}
\alias{gMeasureLeg}
\title{Measure tibia and femur}
\usage{
gMeasureLeg(landmarks, scale)
}
\description{
Scale from pixel to micrometer, and landmarks as outputs of gConnectLeg or gLegLandmarks
}
