% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{GMM}
\alias{GMM}
\title{GMM 'V' using Mclust}
\usage{
GMM(img_val, n = 1:9, keep_max = T, viz = F, n_noise_sd = 0)
}
\description{
returns the GMM model of a distribution as a numeric vector img_val and a dataframe of the parameters with possibility
of removing the one with the highest mean. n integer vector of components to test
if n length > 1 : the number of components will be decided using BIC
}
