% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{GMMdRoots}
\alias{GMMdRoots}
\title{Roots of the derivative of a GMM}
\usage{
GMMdRoots(GMM_df, y, viz = F, msg = T)
}
\description{
Give roots of y for the derivative of given GMM parameters as dataframe created with GMM function
}
