% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{GMMquant}
\alias{GMMquant}
\title{Quantile off GMM model dataframe gmm_model_df from GMM function}
\usage{
GMMquant(gmm_model_df, q, viz = FALSE)
}
\arguments{
\item{gmm_model_df}{dataframe ($df) output of GMM()}

\item{q}{numerical in range 0:1. quantile}

\item{viz}{visualization option}
}
\description{
Quantile off GMM model dataframe gmm_model_df from GMM function
}
