% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gModels.R
\name{gNorm}
\alias{gNorm}
\title{Body contour normalization pipeline for LDA models}
\usage{
gNorm(body, ang, resamp)
}
\arguments{
\item{body}{c.img. Segmented body.}

\item{ang}{numerical. Body elongation angle.}

\item{resamp}{positive integer. Resampling size of body contour.}
}
\description{
Body contour normalization pipeline for LDA models
}
