% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPipeline.R
\name{gOneLeg}
\alias{gOneLeg}
\title{Keep only one tibia and one femur column}
\usage{
gOneLeg(df_result)
}
\arguments{
\item{df_result}{result dataframe as outout of gPipeline or gRunPipeline}
}
\description{
Transform result dataframe to only have one tibia and one femur columns
If both legs are present, return mean values. Removes NA lines.
}
