% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gOrientation}
\alias{gOrientation}
\title{Gerris orientation angle}
\usage{
gOrientation(body_coords, intersection_coords, diag = FALSE)
}
\arguments{
\item{body_coords}{list or single dataframe or matrix of body points coordinates}

\item{intersection_coords}{list or single dataframe or matrix of
full individual and dilated body contour intersection coordinates}

\item{diag}{a logical for output analysis diagnostic messages}
}
\description{
Orientation based on body points PCA, then antero-posterior point density, or appendages repartition
}
