% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPipeline.R
\name{gPipeline}
\alias{gPipeline}
\title{Waterstrider image analysis pipeline}
\usage{
gPipeline(
  img_path,
  write_output = T,
  return_df = T,
  single_write = T,
  thresh = 0.8,
  winged_names = c(0, 1),
  auto_scale = T
)
}
\arguments{
\item{img_path}{image path}

\item{write_output}{logical, to save metrics as csv and print image analysis diagnostic plots}

\item{return_df}{logical, to return metrics in R}

\item{single_write}{logical, writing method for gWritePipeline()}

\item{thresh}{numerical value in range 0:1, global gray-value threshold
for the initial binarization of the input image}

\item{winged_names}{vector, two names for not-winged / winged individuals}

\item{auto_scale}{a boolean. If 'TRUE' : red ruler based scale}
}
\description{
Finds individual's size, sex, presence of wing, tibia and femur length
}
