% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gModels.R
\name{gPredict}
\alias{gPredict}
\title{Predict sex and presence of wings of Microvelia longipes}
\usage{
gPredict(body, angle)
}
\arguments{
\item{body}{a c.img (or list of this type). Segmented body.}

\item{angle}{a numerical (or vector of this type). Orientation angle of
corresponding body}
}
\description{
Use body EFA coefficients to predict sex and presence of wings using
trained LDA model
}
