% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_models.R
\name{gPredictLDA}
\alias{gPredictLDA}
\title{Predict factor using LDA model on harmonic fourier principal components}
\usage{
gPredictLDA(LDA_model, hPCA_scores)
}
\arguments{
\item{LDA_model}{LDA model created with MASS, trained on PCA output}

\item{hPCA_scores}{harmonic fourier principal components as output of prcomp}
}
\description{
Predict factor using LDA model on harmonic fourier principal components
}
