% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gModels.R
\name{gPredictWing}
\alias{gPredictWing}
\title{Predict presence of wing using LDAs}
\usage{
gPredictWing(PCA_scores, sex_class, LDA_f, LDA_m, wing_names = c(1, 2))
}
\arguments{
\item{PCA_scores}{matrix or dataframe with PC scores for individuals}

\item{sex_class}{factor with F & M levels of individuals}

\item{LDA_f}{MASS lda class object}

\item{LDA_m}{MASS lda class object}

\item{wing_names}{a vector of length 2 indicating how the presence of wings should be written}
}
\description{
Predict presence of wing using LDAs
}
