% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grayThresh}
\alias{grayThresh}
\title{Load and binarize image}
\usage{
grayThresh(img, threshold)
}
\arguments{
\item{img}{a cimg.}

\item{threshold}{a numerical in range 0:1 for binarization threshold}
}
\description{
Load image from given path, negative filter, illumination correction filter, binarization based on threshold
}
