% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gSplitLegMeasures}
\alias{gSplitLegMeasures}
\title{Re-format output of gMeasureLeg for easy inclusion in dataframe}
\usage{
gSplitLegMeasures(leg_list)
}
\description{
Re-format output of gMeasureLeg for easy inclusion in dataframe
}
