% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gSubBody}
\alias{gSubBody}
\title{Substract points from image (also invert grayscale)}
\usage{
gSubBody(base_body, seg_body)
}
\arguments{
\item{base_body}{base body cimg. Can be passed as list}

\item{seg_body}{binary body cimg. Can be passed as list}
}
\description{
Substract points from image (also invert grayscale)
}
