% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPipeline.R
\name{gWritePipeline}
\alias{gWritePipeline}
\title{Write waterstriders pipeline outputs in local directory}
\usage{
gWritePipeline(
  img_path,
  i_plots,
  detection_plot,
  df_out,
  dim_img,
  single_write = T
)
}
\arguments{
\item{img_path}{valid path to processed image, used to define writing directory}

\item{i_plots}{list of plots of individuals as created by gPipeline with write_output=T}

\item{detection_plot}{whole image detection plot as created by gPipeline with write_output=T}

\item{df_out}{corresponding dataframe created by gPipeline}

\item{dim_img}{dimensions of base image}

\item{single_write}{method of writing
True : a directory is created for the current image, containing plots and dataframe
False : a single 'out' directory with all plots is created for the current directory
,gMultiPipeline is supposed to take in charge the creation of the dataframe}
}
\description{
Plot and dataframe Writing function for gPipeline, an alternative printing method is
implemented for gMultiPipeline by passing single_write=False
}
