% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{hist_eq}
\alias{hist_eq}
\title{Histogram equalization}
\usage{
hist_eq(img)
}
\arguments{
\item{img}{an imager::cimg}
}
\value{
an imager::cimg
}
\description{
Flatten histogram by replacing the pixel value of an image by their rank.
}
