% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{img_to_df}
\alias{img_to_df}
\title{DataFrame conversion}
\usage{
img_to_df(img)
}
\arguments{
\item{img}{an imager::cimg}
}
\value{
a data.frame
}
\description{
Convert image to dataframe and expose color channel.
}
