% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{intersect_borders}
\alias{intersect_borders}
\title{Intersect borders}
\usage{
intersect_borders(pixset, img)
}
\arguments{
\item{pixset}{an imager::pixset}

\item{img}{an imager::cimg}
}
\value{
a vector of boolean
}
\description{
Test whether a pixset intersect an image borders.
}
