% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{invert_grayscale}
\alias{invert_grayscale}
\title{Invert grayscale image}
\usage{
invert_grayscale(img)
}
\arguments{
\item{img}{an imager::cimg}
}
\value{
an imager::cimg
}
\description{
Invert grayscale image
}
