% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gModels.R
\name{itxFeature}
\alias{itxFeature}
\title{Body-limbs intersection feature builder for orientation prediction model.}
\usage{
itxFeature(dil, int, ang)
}
\arguments{
\item{dil}{list of dilated contour}

\item{int}{list of limbs intersection index on dilcont}

\item{ang}{list of body elongation axis angle}
}
\description{
Creates feature vector with: Ratio of intersection points in 3 equal-sized
bins along dilated body elongation axis, and number of continuous
intersection sequences along dilated body contour.
}
