% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPipeline.R
\name{legText}
\alias{legText}
\title{Quick leg segmentation summary}
\usage{
legText(df)
}
\arguments{
\item{df}{A dataframe with df_out format from gPipeline}
}
\description{
Femur segmentation summary for quick diagnostic in the console
}
