% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_models.R
\name{loadLDA}
\alias{loadLDA}
\title{Load LDA model for package}
\usage{
loadLDA(model_name)
}
\arguments{
\item{model_name}{a string corresponding to the name of the model}
}
\description{
Load LDA model for package
}
