% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gModels.R
\name{loadModel}
\alias{loadModel}
\title{Load LDA model for package}
\usage{
loadModel(model_name)
}
\arguments{
\item{model_name}{a string corresponding to the name of the model}
}
\value{
A fitted LDA or PCA model
}
\description{
Load LDA model for package
}
