% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{major_axis_angle}
\alias{major_axis_angle}
\title{Major axis angle}
\usage{
major_axis_angle(outline)
}
\arguments{
\item{outline}{matrix or dataframe of points coordinates}
}
\description{
Find major axis angle using PCA on outline
}
\keyword{internal}
