% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{make_disc_kernel}
\alias{make_disc_kernel}
\title{Make Disc Kernel}
\usage{
make_disc_kernel(size)
}
\arguments{
\item{size}{numeric value for disc size, will be rounded to the next odd number*}
}
\description{
Make Filled Disc Kernel matrix for morphological dilations or erosions
}
\keyword{internal}
