% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gModels.R
\name{normBody}
\alias{normBody}
\title{Body contour and orientation}
\usage{
normBody(img_bin, ori_angle)
}
\arguments{
\item{img_bin}{c.img or list of c.img of waterstrider body image(s)}

\item{ori_angle}{angle, vector of angles or list of angles of the segmented bodies}
}
\description{
Orient body contours extracted from image using provided angles then define consistent
contour starting point as leftmost point
}
