% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{overlapPoints}
\alias{overlapPoints}
\title{With 2 lists of discrete 2D points : which points from set_a are in set_b}
\usage{
overlapPoints(set_a, set_b, coords = T, index = T)
}
\description{
With 2 lists of discrete 2D points : which points from set_a are in set_b
}
