% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{pipelineScale}
\alias{pipelineScale}
\title{Scale method selecter for gPipeline}
\usage{
pipelineScale(img, auto_scale, red_thresh = 0.05)
}
\arguments{
\item{img}{c.img or pixset}

\item{auto_scale}{boolean for redRulerScale or manual}

\item{red_thresh}{required difference in intensity between red component
and other components to detect the red ruler}
}
\description{
Scale method selecter for gPipeline
}
