% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUtils.R
\name{raster_channel}
\alias{raster_channel}
\title{Raster channel}
\usage{
raster_channel(img, ch = NULL)
}
\arguments{
\item{img}{an imager::cimg}

\item{ch}{character, specify the image channel, one of "R", "G", "B".}
}
\value{
a ggplot2 graph
}
\description{
Raster ggplot of an image channel.
}
