% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{recropBodies}
\alias{recropBodies}
\title{Crop individual's body using crop coordinates given by gCrop}
\usage{
recropBodies(body_coords, base_img, crop_coords)
}
\description{
base_img : a single c.img / body_coords : list of body coordinates / crop_coords : list of 2x2 crop matrices
}
