% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{recropImg}
\alias{recropImg}
\title{Crop image using gCrop(...)$crop_coords}
\usage{
recropImg(img, crop_coords)
}
\arguments{
\item{img}{a cimg object}

\item{crop_coords}{crop coordinates as output of gCrop(...)$crop_coords}
}
\description{
Crop image using gCrop(...)$crop_coords
}
