% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{redRulerScale}
\alias{redRulerScale}
\title{Get scale using detection of red millimeter paper}
\usage{
redRulerScale(
  img,
  red_thresh,
  alpha_confidence = 0.05,
  viz = FALSE,
  msg = TRUE
)
}
\arguments{
\item{img}{c.img or pixset. base image}

\item{red_thresh}{required difference in intensity between red component
and other components to detect the red ruler}

\item{alpha_confidence}{a numerical value in range 0:1.
alpha level for confidence interval}

\item{viz}{logical. visualization option}

\item{msg}{logical. console message option}
}
\description{
Detect ruler as biggest red label
}
