% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{rotation_angle}
\alias{rotation_angle}
\title{Rotation angle}
\usage{
rotation_angle(pixset, size = 6L)
}
\arguments{
\item{pixset}{an imager::pixset}

\item{size}{an integer, the morphological opening factor (optional)}
}
\value{
a real
}
\description{
Get the rotation angle to align the abdomen along the x-axis. The function
fits an ellipsoid hull around the pixset to derive the rotation angle.
}
