% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUtils.R
\name{sample_histogram}
\alias{sample_histogram}
\title{Sample pixel values from histogram}
\usage{
sample_histogram(obj, ratio = 0.02)
}
\arguments{
\item{obj}{an array-like object (img, pixset)}

\item{ratio}{a real, the sampling ratio.}
}
\value{
an integer vector
}
\description{
Sample pixel values from histogram
}
