% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{scalar}
\alias{scalar}
\title{Scalar product}
\usage{
scalar(u, v)
}
\arguments{
\item{u}{a numerical vector of length 2}

\item{v}{a numerical vector of length 2}
}
\description{
Scalar product
}
\keyword{internal}
