% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gModels.R
\name{scoresEFA}
\alias{scoresEFA}
\title{PCA scores of EFA of oriented contours from binary image}
\usage{
scoresEFA(img_bin, ori_angle = rep(0, length(img_bin)), nb_h = 7, viz = FALSE)
}
\arguments{
\item{img_bin}{a list of binary image or pixset with a single white object}

\item{ori_angle}{a list of orientation angles of provided shapes}

\item{nb_h}{a numerical integer. number of elliptuc fourier harmonics to compute}

\item{viz}{visualization option, additionally returns 3 plots}
}
\description{
PCA scores of EFA of oriented contours from binary image
}
