% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gModels.R
\name{scoresEFA2}
\alias{scoresEFA2}
\title{Momocs EFA from waterstrider body image}
\usage{
scoresEFA2(ori_cont, nb_h = 7, viz = FALSE, resamp = 100)
}
\arguments{
\item{ori_cont}{contour or list of contours (as 2D numeric dataframe of matrix) of waterstrider body image(s)}

\item{nb_h}{number of harmonics to compute}

\item{viz}{logical for plot of alignment, fourier space and PCA}

\item{resamp}{number of points to resample on each contour}
}
\description{
Momocs EFA from waterstrider body image
}
