% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{simple_imager_contour}
\alias{simple_imager_contour}
\title{Imager::contours but with simplified output format}
\usage{
simple_imager_contour(pixset)
}
\description{
Imager::contours but with simplified output format
}
