% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{spikePlateau}
\alias{spikePlateau}
\title{Add pike to plateaus of a numeric vector}
\usage{
spikePlateau(v, increment = 1e-05)
}
\arguments{
\item{v}{a numeric vector}

\item{increment}{increment coefficient "size of pyramid steps"}
}
\description{
Add pike to plateaus of a numeric vector
}
