% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{square_stencil}
\alias{square_stencil}
\title{Square stencil}
\usage{
square_stencil(centroid, pad, obj)
}
\arguments{
\item{centroid}{an integer vector, the stencil centroid.}

\item{pad}{an integer, the padding around the stencil centroid.}

\item{obj}{an imager::cimg or an imager::pixset}
}
\value{
a data.frame
}
\description{
Define a square stencil within the boundaries of an image cropping points
that are out of bounds.
}
