% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{tilesUnitDist}
\alias{tilesUnitDist}
\title{4 Nearest neighbor distance for set of points}
\usage{
tilesUnitDist(tiles_centroid, alpha = 0.05, viz = FALSE)
}
\arguments{
\item{tiles_centroid}{tiles centroid as output of redRulerTiles()}

\item{alpha}{confidence interval level}

\item{viz}{a boolean. visualization option}
}
\description{
4 Nearest neighbor distance for set of points
}
