% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{vizLegLandmarks}
\alias{vizLegLandmarks}
\title{Visualization function for gLegLandmarks()}
\usage{
vizLegLandmarks(
  cont_inser,
  inser_thresh,
  dist_peaks_str,
  dist_peaks_rev,
  dist,
  knee,
  ankle,
  knee_pts,
  ankle_pts,
  peaks_inser
)
}
\arguments{
\item{cont_inser}{cicular contour}

\item{inser_thresh}{threshold to remove angular points too close to the leg insertion point}

\item{dist_peaks_str}{distance of each peak to the insertion point}

\item{dist_peaks_rev}{counterclockwise distance of each peak to the insertion point}

\item{dist}{distance between pairs of points}

\item{knee}{knee point}

\item{ankle}{ankle point}

\item{knee_pts}{knee joint points}

\item{ankle_pts}{ankle joint points}

\item{peaks_inser}{names to track index in contour}
}
\description{
Plot 1: Inflexion points to insertion point distance + insertion threshold
PLot 2: Contour reordered around insertion point + inflexion points
}
