% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherUtils.R
\name{white_pix_bool}
\alias{white_pix_bool}
\title{Get white pixels coordinates of a pixset for boolean files}
\usage{
white_pix_bool(pixset)
}
\description{
Get white pixels coordinates of a pixset for boolean files
}
