% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{binaryLoad}
\alias{binaryLoad}
\title{Load and binarize image}
\usage{
binaryLoad(img_path, threshold)
}
\arguments{
\item{img_path}{a string. path to the base image}

\item{threshold}{a numerical in range 0:1 for binarization threshold}
}
\description{
Load image from given path, negative filter, illumination correction filter, binarization based on threshold
}
