% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{bodyLength}
\alias{bodyLength}
\title{Body length from body points}
\usage{
bodyLength(body_pts, viz = FALSE, return_ext = FALSE)
}
\arguments{
\item{body_pts}{binary object as point cloud of even density}

\item{viz}{logical. visualization option}

\item{return_ext}{logical. output format, if TRUE, also returns points between which length was measured}
}
\description{
Body length as distance between intersections of body outline with line passing
through the centroid of the body and minimizing lateral point density difference
along body elongation axis defined using PCA
}
