% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{bodyLength}
\alias{bodyLength}
\title{Body length from body points}
\usage{
bodyLength(body_points)
}
\arguments{
\item{body_points}{a dataframe or matrix of body points coordinates}
}
\description{
Body length from body points
}
