% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{cleanBodyShape}
\alias{cleanBodyShape}
\title{Clean body shape as xy coordinates using morphological closing following by opening}
\usage{
cleanBodyShape(body_lab_points, kernel_size = 3, as_coords = T)
}
\arguments{
\item{body_lab_points}{body points as numerical matrix or dataframe}

\item{kernel_size}{numerical integer. closing and opening kernel size in pixels}

\item{as_coords}{logical. to return}
}
\description{
Square closing then closing of size kernel_size on body coordinates body_lab_points
Returns either coordinates as_coord, or an image c.img
}
