% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{cleanBodyShape}
\alias{cleanBodyShape}
\title{Clean body shape as xy coordinates using morphological closing following by opening}
\usage{
cleanBodyShape(body_lab_points, kernel_size = 3, as_coords = T)
}
\description{
Square closing then closing of size kernel_size on body coordinates body_lab_points
Returns either coordinates as_coord, or an image c.img
}
