% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{contourAngles}
\alias{contourAngles}
\title{Contour angle for each point given a search window = index of the next value to check}
\usage{
contourAngles(coords, search_w = 5)
}
\arguments{
\item{coords}{dataframe or matrix of contour coordinates}

\item{search_w}{numerical integer. Index of the next value to check in the contour}
}
\description{
Contour angle for each point given a search window = index of the next value to check
}
