% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{coordsToLinear}
\alias{coordsToLinear}
\title{Convert xy coordinates as linear index in given c.img}
\usage{
coordsToLinear(coords, img)
}
\arguments{
\item{coords}{matrix or dataframe of numerical coordinates}

\item{img}{a c.img}
}
\description{
Convert xy coordinates as linear index in given c.img
}
