% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{f_prime}
\alias{f_prime}
\title{Accesory function for GMMdRoots : creates derivative function of GMM}
\usage{
f_prime(x, GMM_df)
}
\arguments{
\item{x}{A numeric vector. The input values for which
the derivative of the Gaussian Mixture Model is to be computed.}

\item{GMM_df}{dataframe ($df) output of GMM()}
}
\description{
Accesory function for GMMdRoots : creates derivative function of GMM
}
\keyword{internal}
