% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gCrop}
\alias{gCrop}
\title{Crop individual gerris}
\usage{
gCrop(base_img, centroids, sizes, viz = FALSE, factor = 3.5)
}
\arguments{
\item{base_img}{c.img. Base image suitable for waterstrider pipeline}

\item{centroids}{list of xy vectors. Body centroids coordinates of individuals}

\item{sizes}{list or vector. Body length of individuals}

\item{viz}{logical. visualization option}

\item{factor}{numerical. Numerical value for crop size relative to individual's body length}
}
\description{
Square crop around individual gerris centroids from image base_img with a diameter of factor*size
Base image must be imported with imager package, centroids and sizes must be lists of same length
Returns a list of inverted grayscale cropped images
}
