% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gDetectionPlot}
\alias{gDetectionPlot}
\title{Plot gerris position and id with 1cm scale on original image}
\usage{
gDetectionPlot(base_img, scale, x, y)
}
\arguments{
\item{base_img}{c.img or pixset of base image}

\item{scale}{mean pixel amount for 1 micron}

\item{x, y}{ordered position of individuals}
}
\description{
Plot gerris position and id with 1cm scale on original image
}
