% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gFastSeg}
\alias{gFastSeg}
\title{Fast segmentation with size filter}
\usage{
gFastSeg(img_bin, min = 300, max = 1500)
}
\arguments{
\item{img_bin}{a pixset}

\item{min}{minimum number of pixels to keep label}

\item{max}{maximum number of pixels to keep label}
}
\description{
Given a binary image as pixset, returns list of coordinates of every label with pixels in the min-max range
}
