% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gGerrisPlot}
\alias{gGerrisPlot}
\title{Individual plot for gPipeline}
\usage{
gGerrisPlot(
  i,
  full,
  body,
  cen,
  dilcont,
  ang,
  legs,
  leg_lm,
  leg_size,
  inser,
  clean_base_path
)
}
\description{
Individual plot for gPipeline
}
