% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gLegLandmarksLoop}
\alias{gLegLandmarksLoop}
\title{Loop logic for leg landmarking in gPipeline}
\usage{
gLegLandmarksLoop(leg_coords, insertion, ...)
}
\arguments{
\item{leg_coords}{numerical matrix or dataframe or list of those types.
leg points coordinates}

\item{insertion}{numerical matrix or dataframe or list of those types.
insertion points coordinates}

\item{...}{arguments to be passed to gLegLandmarks()}
}
\description{
Allows handling of more than one leg when provided with a specific input structure :
Use case 1 - One individual : assumes named list of two elements "right" and "left"
Use case 2 - Multiple individuals : assumes list of individuals as described in case 1
}
