% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gMeasureLeg}
\alias{gMeasureLeg}
\title{Measure tibia and femur}
\usage{
gMeasureLeg(landmarks, scale)
}
\arguments{
\item{landmarks}{landmarks as output of gLegLandmarks()}

\item{scale}{a numerical value for conversion from pixels to micrometers}
}
\description{
Scale from pixel to micrometer, and landmarks as outputs of gConnectLeg or gLegLandmarks
}
