% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{GMM}
\alias{GMM}
\title{GMM 'V' using Mclust}
\usage{
GMM(img_val, n = 1:9, keep_max = T, viz = F, n_noise_sd = 0)
}
\arguments{
\item{img_val}{numerical gray values of an image}

\item{n}{numerical integer in range 1:9. number 'G' of components of the GMM}

\item{keep_max}{logical. to remove rightmost component and consequently
adjust weights}

\item{viz}{visualization option}

\item{n_noise_sd}{numerical. standard deviation of normal noise to add
to the data to help breaking plateaus in he data, not compatible with findPeaks()}
}
\description{
returns the GMM model of a distribution as a numeric vector img_val and a dataframe of the parameters with possibility
of removing the one with the highest mean. n integer vector of components to test
if n length > 1 : the number of components will be decided using BIC
}
