% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{GMMdRoots}
\alias{GMMdRoots}
\title{Roots of the derivative of a GMM}
\usage{
GMMdRoots(GMM_df, y, viz = FALSE, msg = TRUE)
}
\arguments{
\item{GMM_df}{dataframe ($df) output of GMM()}

\item{y}{numerical value to find the roots of}

\item{viz}{visualization option}

\item{msg}{onsole pessage option}
}
\description{
Give roots of y for the derivative of given GMM parameters as dataframe created with GMM function
}
