% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{GMMquant}
\alias{GMMquant}
\title{Quantile off GMM model dataframe gmm_model_df from GMM function}
\usage{
GMMquant(gmm_model_df, q, viz = F)
}
\description{
Quantile off GMM model dataframe gmm_model_df from GMM function
}
