% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPipeline.R
\name{gMultiPipeline}
\alias{gMultiPipeline}
\title{Pipeline iteration for multiple images}
\usage{
gMultiPipeline(img_path_list, return_df)
}
\arguments{
\item{img_path_list}{a vector or list or valid image paths}

\item{return_df}{logical, to return the concatenated gPipeline dataframe}
}
\description{
gPipeline iteration for multiple images including specific writing option,
and appropriate console messages
}
