% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gOrientation}
\alias{gOrientation}
\title{Gerris orientation angle}
\usage{
gOrientation(body_coords, intersection_coords, diag = F)
}
\description{
Orientation based on body points PCA, then antero-posterior point density, or appendages repartition
}
