% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{gPipeline}
\alias{gPipeline}
\title{The main thing... for now}
\usage{
gPipeline(img_path, write_df = T, write_plots = T, df_output = T)
}
\arguments{
\item{img_path}{image path or list of image paths}

\item{write_df}{a boolean, to save metrics as csv}

\item{write_plots}{a boolean, to save image analysis diagnostic plots}

\item{df_output}{a boolean, to return metrics in R}
}
\description{
The main thing... for now
}
