% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{openGrDevice}
\alias{openGrDevice}
\title{Open a Graphics Device Depending on OS}
\usage{
openGrDevice()
}
\value{
\code{TRUE} if the graphics device was successfully opened, \code{FALSE} otherwise.
}
\description{
Opens an interactive graphics device suitable for the current operating system.
\itemize{
\item On \strong{macOS}, uses \code{quartz()}
\item On \strong{Linux} and \strong{Windows}, uses \code{X11()}
}
}
\examples{
\dontrun{
if (interactive()) {
  if (openGrDevice()) {
    plot(1:10)
  }
}
}
}
