% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{overlapPoints}
\alias{overlapPoints}
\title{With 2 lists of discrete 2D points : which points from set_a are in set_b}
\usage{
overlapPoints(set_a, set_b, coords = TRUE, index = TRUE)
}
\arguments{
\item{set_a}{a numerical matrix of 2D points}

\item{set_b}{a numerical matrix of 2D points}

\item{coords}{logical. to return numerical points coordinates}

\item{index}{logical. to return point's index}
}
\description{
With 2 lists of discrete 2D points : which points from set_a are in set_b
}
