% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{pipelineScale}
\alias{pipelineScale}
\title{Scale method selecter for gPipeline}
\usage{
pipelineScale(img, auto_scale)
}
\arguments{
\item{img}{c.img or pixset}

\item{auto_scale}{boolean for redRulerScale or manual}
}
\description{
Scale method selecter for gPipeline
}
