% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFun.R
\name{recropBodies}
\alias{recropBodies}
\title{Crop individual's body using crop coordinates given by gCrop}
\usage{
recropBodies(body_coords, base_img, crop_coords)
}
\arguments{
\item{body_coords}{a list of body coordinates}

\item{base_img}{c.img. Base image suitable for waterstrider pipeline}

\item{crop_coords}{a list of 2x2 crop matrices}
}
\description{
Crop individual's body using crop coordinates given by gCrop
}
