% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{redRulerScale}
\alias{redRulerScale}
\title{Get scale using detection of red millimeter paper}
\usage{
redRulerScale(
  img,
  red_thresh = 0.05,
  confidence_interval = 0.95,
  viz = F,
  msg = T
)
}
\description{
Detect ruler as biggest red label,
}
