% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{redRulerScale}
\alias{redRulerScale}
\title{Get scale using detection of red millimeter paper}
\usage{
redRulerScale(
  img,
  red_thresh = 0.05,
  confidence_interval = 0.95,
  viz = F,
  msg = T
)
}
\arguments{
\item{img}{c.img or pixset. base image}

\item{red_thresh}{required difference in intensity between red component
and other components to detect the red ruler}

\item{confidence_interval}{a numerical value in range 0:1.
confidence interval value to build the error margin from}

\item{viz}{logical. visualization option}

\item{msg}{logical. console message option}
}
\description{
Detect ruler as biggest red label,
}
